requirements.txt
firebase-admin==6.1.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
SQLAlchemy==1.4.47
requests==2.28.2
PyJWT~=2.6.0
DateTime~=5.1
numpy~=1.23.5
pandas~=1.5.3
tensorflow~=2.12.0
scikit-learn~=1.2.2
matplotlib~=3.7.1
streamlit~=1.21.0
yfinance~=0.2.17
prophet~=1.1.2
plotly~=5.14.1